refactor: organize core integration tests by service
Organizes the core integration tests to mirror the structure in
node.js and web introduced in the prior two commits.
