+++ packages/common-integration-tests/src/cache/update-ttl.ts
@@ -3,7 +3,7 @@ import {
   ValidateCacheProps,
   ItBehavesLikeItValidatesCacheName,
   expectWithMessage,
-} from './common-int-test-utils';
+} from '../common-int-test-utils';
 import {ICacheClient} from '@gomomento/sdk-core/dist/src/internal/clients/cache';
 import {
   CacheUpdateTtl,
